requirements.txt

fastapi
uvicorn
python-dotenv
openai
pinecone
azure-storage-blob
PyPDF2
python-docx
